support for CRLN delimited events
does not handle CR, despite what the spec says.

anything but LN is stupid. CR/CRLN is the only tricky part of the whole spec.
worse, it effectively prevents the message data from ever containing a
CR. multiple data messages can be sent for LN, and the reader joins them
back up, but they are always joined back up with 'LN' (as there's no way
to know the original separator).